rasa/api/requirements.txt
# Flask API requirements
Flask==2.3.3
Flask-CORS==4.0.0

# Security
PyJWT==2.8.0
redis==4.6.0

# HTTP requests
requests==2.31.0

# Production server
gunicorn==21.2.0